requirements.txt
dash==2.18.2
plotly==5.24.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
bcrypt==4.2.1
dash-bootstrap-components==1.6.0
gunicorn==23.0.0
unidecode==1.3.8
